Rename game to Niwatori
Niwatori (鶏) is a Japanese word for "chicken" (the animal, not
poultry). I chose this name to pay homage to Sokoban (倉庫番), which my
game plays similarly to.

My game was inspired by a game I made in college for a game project
course. I do not remember what inspired that game--I was not familiar
with Sokoban until I started building Niwatori for js13k 2024.